Fix float value conversion oversight
Fixes #12461

+++ community/cypher/interpreted-runtime/src/main/scala/org/neo4j/cypher/internal/runtime/interpreted/ValueConversion.scala
@@ -109,7 +109,7 @@ object ValueConversion {
     case s: String => Values.utf8Value(s)
     case c: Char => Values.charValue(c)
     case d: Double => Values.doubleValue(d)
-    case f: Float => Values.doubleValue(f)
+    case f: Float => Values.floatValue(f)
     case n: Number => Values.longValue(n.longValue())
     case b: Boolean => Values.booleanValue(b)
     case n: Node => ValueUtils.fromNodeEntity(n)
